Refactor Interactive Window Tests
- Create a base class for Interactive Window Tests
- Create a class for tests of clipboard related operations
