+++ src/InteractiveWindow/EditorTest/InteractiveWindowHistoryTests.cs
@@ -7,22 +7,20 @@ using Xunit;
 
 namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
 {
-    public class InteractiveWindowHistoryTests : IDisposable
+    public class InteractiveWindowHistoryTests : InteractiveWindowTestBase
     {
         #region Helpers
 
-        private readonly InteractiveWindowTestHost _testHost;
-        private readonly IInteractiveWindow _window;
-        private readonly IInteractiveWindowOperations _operations;
+        private readonly InteractiveWindowTestHost _testHost;    
 
         public InteractiveWindowHistoryTests()
         {
-            _testHost = new InteractiveWindowTestHost();
-            _window = _testHost.Window;
-            _operations = _window.Operations;
+            _testHost = new InteractiveWindowTestHost();  
         }
 
-        void IDisposable.Dispose()
+        internal override InteractiveWindowTestHost TestHost => _testHost;
+
+        public override void Dispose()
         {
             _testHost.Dispose();
         }
@@ -32,9 +30,9 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
         /// </summary>
         private void SetActiveCode(string text)
         {
-            using (var edit = _window.CurrentLanguageBuffer.CreateEdit(EditOptions.None, reiteratedVersionNumber: null, editTag: null))
+            using (var edit = Window.CurrentLanguageBuffer.CreateEdit(EditOptions.None, reiteratedVersionNumber: null, editTag: null))
             {
-                edit.Replace(new Span(0, _window.CurrentLanguageBuffer.CurrentSnapshot.Length), text);
+                edit.Replace(new Span(0, Window.CurrentLanguageBuffer.CurrentSnapshot.Length), text);
                 edit.Apply();
             }
         }
@@ -49,29 +47,64 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
 
         private void InsertAndExecuteInput(string input)
         {
-            _window.InsertCode(input);
+            Window.InsertCode(input);
             AssertCurrentSubmission(input);
             ExecuteInput();
         }
 
         private void ExecuteInput()
         {
-            ((InteractiveWindow)_window).ExecuteInputAsync().PumpingWait();
+            ((InteractiveWindow)Window).ExecuteInputAsync().PumpingWait();
         }
 
         private void AssertCurrentSubmission(string expected)
         {
-            Assert.Equal(expected, _window.CurrentLanguageBuffer.CurrentSnapshot.GetText());
+            Assert.Equal(expected, Window.CurrentLanguageBuffer.CurrentSnapshot.GetText());
         }
 
-        #endregion Helpers
+        #endregion Helpers    
+
+        [WpfFact]
+        public void CallClearHistoryOnNonUIThread()
+        {
+            Window.AddInput("1"); // Need a history entry.
+            TaskRun(() => Window.Operations.ClearHistory()).PumpingWait();
+        }
+
+        [WpfFact]
+        public void CallHistoryNextOnNonUIThread()
+        {
+            Window.AddInput("1"); // Need a history entry.
+            TaskRun(() => Window.Operations.HistoryNext()).PumpingWait();
+        }
+
+        [WpfFact]
+        public void CallHistoryPreviousOnNonUIThread()
+        {
+            Window.AddInput("1"); // Need a history entry.
+            TaskRun(() => Window.Operations.HistoryPrevious()).PumpingWait();
+        }
+
+        [WpfFact]
+        public void CallHistorySearchNextOnNonUIThread()
+        {
+            Window.AddInput("1"); // Need a history entry.
+            TaskRun(() => Window.Operations.HistorySearchNext()).PumpingWait();
+        }
+
+        [WpfFact]
+        public void CallHistorySearchPreviousOnNonUIThread()
+        {
+            Window.AddInput("1"); // Need a history entry.
+            TaskRun(() => Window.Operations.HistorySearchPrevious()).PumpingWait();
+        }
 
         [WpfFact]
         public void CheckHistoryPrevious()
         {
             const string inputString = "1 ";
             InsertAndExecuteInput(inputString);
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString);
         }
 
@@ -84,12 +117,12 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             InsertAndExecuteInput(inputString1);
             InsertAndExecuteInput(inputString2);
 
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString2);
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString1);
             //this up should not be circular
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString1);
         }
 
@@ -105,23 +138,23 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             InsertAndExecuteInput(inputString2);
             InsertAndExecuteInput(inputString3);
 
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString3);
 
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString2);
 
             ExecuteInput();
 
             //history navigation should start from the last history pointer
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString2);
 
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString1);
 
             //has reached the top, no change
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString1);
         }
 
@@ -136,10 +169,10 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             InsertAndExecuteInput(inputString1);
             InsertAndExecuteInput(inputString2);
 
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString2);
 
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString1);
 
             SetActiveCode(inputString3);
@@ -147,17 +180,17 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             ExecuteInput();
 
             //History pointer should be reset. Previous should now bring up last entry
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString3);
 
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString2);
 
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString1);
 
             //has reached the top, no change
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString1);
         }
 
@@ -173,20 +206,20 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
 
             //Next should do nothing as history pointer is uninitialized and there is
             //no next entry. Buffer should be empty
-            _operations.HistoryNext();
+            Window.Operations.HistoryNext();
             AssertCurrentSubmission(empty);
 
             //Go back once entry
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString2);
 
             //Go fwd one entry - should do nothing as history pointer is at last entry
             //buffer should have same value as before
-            _operations.HistoryNext();
+            Window.Operations.HistoryNext();
             AssertCurrentSubmission(inputString2);
 
             //Next should again do nothing as it is the last item, bufer should have the same value
-            _operations.HistoryNext();
+            Window.Operations.HistoryNext();
             AssertCurrentSubmission(inputString2);
 
             //This is to make sure the window doesn't crash
@@ -206,25 +239,25 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             InsertAndExecuteInput(inputString2);
             InsertAndExecuteInput(inputString3);
 
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString3);
 
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString2);
 
             //submit inputString2 again. Should be added at the end of history
             ExecuteInput();
 
             //history navigation should start from the last history pointer
-            _operations.HistoryNext();
+            Window.Operations.HistoryNext();
             AssertCurrentSubmission(inputString3);
 
             //This next should take us to the InputString2 which was resubmitted
-            _operations.HistoryNext();
+            Window.Operations.HistoryNext();
             AssertCurrentSubmission(inputString2);
 
             //has reached the top, no change
-            _operations.HistoryNext();
+            Window.Operations.HistoryNext();
             AssertCurrentSubmission(inputString2);
         }
 
@@ -240,10 +273,10 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             InsertAndExecuteInput(inputString1);
             InsertAndExecuteInput(inputString2);
 
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString2);
 
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString1);
 
             SetActiveCode(inputString3);
@@ -251,10 +284,10 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             ExecuteInput();
 
             //History pointer should be reset. next should do nothing
-            _operations.HistoryNext();
+            Window.Operations.HistoryNext();
             AssertCurrentSubmission(empty);
 
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             AssertCurrentSubmission(inputString3);
         }
 
@@ -271,10 +304,10 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             //Add uncommitted input
             SetActiveCode(uncommittedInput);
             //Navigate history. This should save uncommitted input
-            _operations.HistoryPrevious();
+            Window.Operations.HistoryPrevious();
             //Navigate to next item at the end of history.
             //This should bring back uncommitted input
-            _operations.HistoryNext();
+            Window.Operations.HistoryNext();
             AssertCurrentSubmission(uncommittedInput);
         }
 
@@ -285,9 +318,9 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             const string resetCommand2 = "#reset  ";
             InsertAndExecuteInput(resetCommand1);
             InsertAndExecuteInput(resetCommand2);
-            _operations.HistoryPrevious();  AssertCurrentSubmission(resetCommand2);
-            _operations.HistoryPrevious();  AssertCurrentSubmission(resetCommand1);
-            _operations.HistoryPrevious();  AssertCurrentSubmission(resetCommand1);
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission(resetCommand2);
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission(resetCommand1);
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission(resetCommand1);
         }
 
         [WpfFact]
@@ -295,11 +328,11 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
         {
             InsertAndExecuteInputs("1", "2", "3");
 
-            _operations.HistoryPrevious();  AssertCurrentSubmission("3");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("2");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("1");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("1");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("1");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("3");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("2");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("1");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("1");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("1");
         }
 
         [WpfFact]
@@ -309,16 +342,16 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
 
             SetActiveCode("4");
 
-            _operations.HistoryNext();      AssertCurrentSubmission("4");
-            _operations.HistoryNext();      AssertCurrentSubmission("4");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("3");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("2");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("1");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("1");
-            _operations.HistoryNext();      AssertCurrentSubmission("2");
-            _operations.HistoryNext();      AssertCurrentSubmission("3");
-            _operations.HistoryNext();      AssertCurrentSubmission("4");
-            _operations.HistoryNext();      AssertCurrentSubmission("4");
+            Window.Operations.HistoryNext();      AssertCurrentSubmission("4");
+            Window.Operations.HistoryNext();      AssertCurrentSubmission("4");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("3");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("2");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("1");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("1");
+            Window.Operations.HistoryNext();      AssertCurrentSubmission("2");
+            Window.Operations.HistoryNext();      AssertCurrentSubmission("3");
+            Window.Operations.HistoryNext();      AssertCurrentSubmission("4");
+            Window.Operations.HistoryNext();      AssertCurrentSubmission("4");
         }
 
         [WpfFact]
@@ -326,8 +359,8 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
         {
             InsertAndExecuteInputs("123", "12", "1");
 
-            _operations.HistoryPrevious("4");   AssertCurrentSubmission("");
-            _operations.HistoryPrevious("4");   AssertCurrentSubmission("");
+            Window.Operations.HistoryPrevious("4");   AssertCurrentSubmission("");
+            Window.Operations.HistoryPrevious("4");   AssertCurrentSubmission("");
         }
 
         [WpfFact]
@@ -335,9 +368,9 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
         {
             InsertAndExecuteInputs("123", "12", "1");
 
-            _operations.HistoryPrevious("12");  AssertCurrentSubmission("12"); // Skip over non-matching entry.
-            _operations.HistoryPrevious("12");  AssertCurrentSubmission("123");
-            _operations.HistoryPrevious("12");  AssertCurrentSubmission("123");
+            Window.Operations.HistoryPrevious("12");  AssertCurrentSubmission("12"); // Skip over non-matching entry.
+            Window.Operations.HistoryPrevious("12");  AssertCurrentSubmission("123");
+            Window.Operations.HistoryPrevious("12");  AssertCurrentSubmission("123");
         }
 
         [WpfFact]
@@ -345,9 +378,9 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
         {
             InsertAndExecuteInputs("1", "2", "3");
 
-            _operations.HistoryPrevious("2");   AssertCurrentSubmission("2"); // Skip over non-matching entry.
-            _operations.HistoryPrevious(null);  AssertCurrentSubmission("1"); // Pattern isn't passed, so return to normal iteration.
-            _operations.HistoryPrevious(null);  AssertCurrentSubmission("1");
+            Window.Operations.HistoryPrevious("2");   AssertCurrentSubmission("2"); // Skip over non-matching entry.
+            Window.Operations.HistoryPrevious(null);  AssertCurrentSubmission("1"); // Pattern isn't passed, so return to normal iteration.
+            Window.Operations.HistoryPrevious(null);  AssertCurrentSubmission("1");
         }
 
         [WpfFact]
@@ -355,9 +388,9 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
         {
             InsertAndExecuteInputs("10", "20", "15", "25");
 
-            _operations.HistoryPrevious("1");   AssertCurrentSubmission("15"); // Skip over non-matching entry.
-            _operations.HistoryPrevious("2");   AssertCurrentSubmission("20"); // Skip over non-matching entry.
-            _operations.HistoryPrevious("2");   AssertCurrentSubmission("20");
+            Window.Operations.HistoryPrevious("1");   AssertCurrentSubmission("15"); // Skip over non-matching entry.
+            Window.Operations.HistoryPrevious("2");   AssertCurrentSubmission("20"); // Skip over non-matching entry.
+            Window.Operations.HistoryPrevious("2");   AssertCurrentSubmission("20");
         }
 
         [WpfFact]
@@ -366,13 +399,13 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             InsertAndExecuteInputs("start", "1", "12", "123");
             SetActiveCode("end");
 
-            _operations.HistoryPrevious();  AssertCurrentSubmission("123");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("12");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("1");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("start");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("123");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("12");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("1");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("start");
 
-            _operations.HistoryNext("4");   AssertCurrentSubmission("end");
-            _operations.HistoryNext("4");   AssertCurrentSubmission("end");
+            Window.Operations.HistoryNext("4");   AssertCurrentSubmission("end");
+            Window.Operations.HistoryNext("4");   AssertCurrentSubmission("end");
         }
 
         [WpfFact]
@@ -381,14 +414,14 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             InsertAndExecuteInputs("start", "1", "12", "123");
             SetActiveCode("end");
 
-            _operations.HistoryPrevious();  AssertCurrentSubmission("123");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("12");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("1");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("start");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("123");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("12");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("1");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("start");
 
-            _operations.HistoryNext("12");  AssertCurrentSubmission("12"); // Skip over non-matching entry.
-            _operations.HistoryNext("12");  AssertCurrentSubmission("123");
-            _operations.HistoryNext("12");  AssertCurrentSubmission("end");
+            Window.Operations.HistoryNext("12");  AssertCurrentSubmission("12"); // Skip over non-matching entry.
+            Window.Operations.HistoryNext("12");  AssertCurrentSubmission("123");
+            Window.Operations.HistoryNext("12");  AssertCurrentSubmission("end");
         }
 
         [WpfFact]
@@ -397,14 +430,14 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             InsertAndExecuteInputs("start", "3", "2", "1");
             SetActiveCode("end");
 
-            _operations.HistoryPrevious();  AssertCurrentSubmission("1");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("2");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("3");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("start");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("1");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("2");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("3");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("start");
 
-            _operations.HistoryNext("2");   AssertCurrentSubmission("2"); // Skip over non-matching entry.
-            _operations.HistoryNext(null);  AssertCurrentSubmission("1"); // Pattern isn't passed, so return to normal iteration.
-            _operations.HistoryNext(null);  AssertCurrentSubmission("end");
+            Window.Operations.HistoryNext("2");   AssertCurrentSubmission("2"); // Skip over non-matching entry.
+            Window.Operations.HistoryNext(null);  AssertCurrentSubmission("1"); // Pattern isn't passed, so return to normal iteration.
+            Window.Operations.HistoryNext(null);  AssertCurrentSubmission("end");
         }
 
         [WpfFact]
@@ -413,15 +446,15 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             InsertAndExecuteInputs("start", "25", "15", "20", "10");
             SetActiveCode("end");
 
-            _operations.HistoryPrevious();  AssertCurrentSubmission("10");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("20");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("15");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("25");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("start");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("10");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("20");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("15");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("25");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("start");
 
-            _operations.HistoryNext("1");   AssertCurrentSubmission("15"); // Skip over non-matching entry.
-            _operations.HistoryNext("2");   AssertCurrentSubmission("20"); // Skip over non-matching entry.
-            _operations.HistoryNext("2");   AssertCurrentSubmission("end");
+            Window.Operations.HistoryNext("1");   AssertCurrentSubmission("15"); // Skip over non-matching entry.
+            Window.Operations.HistoryNext("2");   AssertCurrentSubmission("20"); // Skip over non-matching entry.
+            Window.Operations.HistoryNext("2");   AssertCurrentSubmission("end");
         }
 
         [WpfFact]
@@ -430,10 +463,10 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             InsertAndExecuteInputs("123", "12", "1");
 
             // Default search string is empty.
-            _operations.HistorySearchPrevious();    AssertCurrentSubmission("1"); // Pattern is captured before this step.
-            _operations.HistorySearchPrevious();    AssertCurrentSubmission("12");
-            _operations.HistorySearchPrevious();    AssertCurrentSubmission("123");
-            _operations.HistorySearchPrevious();    AssertCurrentSubmission("123");
+            Window.Operations.HistorySearchPrevious();    AssertCurrentSubmission("1"); // Pattern is captured before this step.
+            Window.Operations.HistorySearchPrevious();    AssertCurrentSubmission("12");
+            Window.Operations.HistorySearchPrevious();    AssertCurrentSubmission("123");
+            Window.Operations.HistorySearchPrevious();    AssertCurrentSubmission("123");
         }
 
         [WpfFact]
@@ -442,9 +475,9 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             InsertAndExecuteInputs("123", "12", "1");
             SetActiveCode("12");
 
-            _operations.HistorySearchPrevious();    AssertCurrentSubmission("12"); // Pattern is captured before this step.
-            _operations.HistorySearchPrevious();    AssertCurrentSubmission("123");
-            _operations.HistorySearchPrevious();    AssertCurrentSubmission("123");
+            Window.Operations.HistorySearchPrevious();    AssertCurrentSubmission("12"); // Pattern is captured before this step.
+            Window.Operations.HistorySearchPrevious();    AssertCurrentSubmission("123");
+            Window.Operations.HistorySearchPrevious();    AssertCurrentSubmission("123");
         }
 
         [WpfFact]
@@ -453,14 +486,14 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             InsertAndExecuteInputs("12", "123", "12", "1");
             SetActiveCode("end");
 
-            _operations.HistoryPrevious();  AssertCurrentSubmission("1");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("12");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("123");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("12");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("1");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("12");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("123");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("12");
 
-            _operations.HistorySearchNext();    AssertCurrentSubmission("123"); // Pattern is captured before this step.
-            _operations.HistorySearchNext();    AssertCurrentSubmission("12");
-            _operations.HistorySearchNext();    AssertCurrentSubmission("end");
+            Window.Operations.HistorySearchNext();    AssertCurrentSubmission("123"); // Pattern is captured before this step.
+            Window.Operations.HistorySearchNext();    AssertCurrentSubmission("12");
+            Window.Operations.HistorySearchNext();    AssertCurrentSubmission("end");
         }
 
         [WpfFact]
@@ -468,13 +501,13 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
         {
             InsertAndExecuteInputs("200", "100", "30", "20", "10", "2", "1");
 
-            _operations.HistoryPrevious();          AssertCurrentSubmission("1");
-            _operations.HistorySearchPrevious();    AssertCurrentSubmission("10"); // Pattern is captured before this step.
-            _operations.HistoryPrevious();          AssertCurrentSubmission("20"); // NB: Doesn't match pattern.
-            _operations.HistorySearchPrevious();    AssertCurrentSubmission("100"); // NB: Reuses existing pattern.
-            _operations.HistorySearchPrevious();    AssertCurrentSubmission("100");
-            _operations.HistoryPrevious();          AssertCurrentSubmission("200");
-            _operations.HistorySearchPrevious();    AssertCurrentSubmission("200"); // No-op results in non-matching history entry after SearchPrevious.
+            Window.Operations.HistoryPrevious();          AssertCurrentSubmission("1");
+            Window.Operations.HistorySearchPrevious();    AssertCurrentSubmission("10"); // Pattern is captured before this step.
+            Window.Operations.HistoryPrevious();          AssertCurrentSubmission("20"); // NB: Doesn't match pattern.
+            Window.Operations.HistorySearchPrevious();    AssertCurrentSubmission("100"); // NB: Reuses existing pattern.
+            Window.Operations.HistorySearchPrevious();    AssertCurrentSubmission("100");
+            Window.Operations.HistoryPrevious();          AssertCurrentSubmission("200");
+            Window.Operations.HistorySearchPrevious();    AssertCurrentSubmission("200"); // No-op results in non-matching history entry after SearchPrevious.
         }
 
         [WpfFact]
@@ -482,13 +515,13 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
         {
             InsertAndExecuteInputs("200", "100", "30", "20", "10", "2", "1");
 
-            _operations.HistoryPrevious();          AssertCurrentSubmission("1");
-            _operations.HistorySearchPrevious();    AssertCurrentSubmission("10"); // Pattern is captured before this step.
-            _operations.HistoryPrevious("2");       AssertCurrentSubmission("20"); // NB: Doesn't match pattern.
-            _operations.HistorySearchPrevious();    AssertCurrentSubmission("100"); // NB: Reuses existing pattern.
-            _operations.HistorySearchPrevious();    AssertCurrentSubmission("100");
-            _operations.HistoryPrevious("2");       AssertCurrentSubmission("200");
-            _operations.HistorySearchPrevious();    AssertCurrentSubmission("200"); // No-op results in non-matching history entry after SearchPrevious.
+            Window.Operations.HistoryPrevious();          AssertCurrentSubmission("1");
+            Window.Operations.HistorySearchPrevious();    AssertCurrentSubmission("10"); // Pattern is captured before this step.
+            Window.Operations.HistoryPrevious("2");       AssertCurrentSubmission("20"); // NB: Doesn't match pattern.
+            Window.Operations.HistorySearchPrevious();    AssertCurrentSubmission("100"); // NB: Reuses existing pattern.
+            Window.Operations.HistorySearchPrevious();    AssertCurrentSubmission("100");
+            Window.Operations.HistoryPrevious("2");       AssertCurrentSubmission("200");
+            Window.Operations.HistorySearchPrevious();    AssertCurrentSubmission("200"); // No-op results in non-matching history entry after SearchPrevious.
         }
 
         [WpfFact]
@@ -497,19 +530,19 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             InsertAndExecuteInputs("1", "2", "10", "20", "30", "100", "200");
             SetActiveCode("4");
 
-            _operations.HistoryPrevious();  AssertCurrentSubmission("200");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("100");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("30");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("20");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("10");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("2");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("1");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("200");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("100");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("30");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("20");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("10");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("2");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("1");
 
-            _operations.HistorySearchNext();    AssertCurrentSubmission("10"); // Pattern is captured before this step.
-            _operations.HistoryNext();          AssertCurrentSubmission("20"); // NB: Doesn't match pattern.
-            _operations.HistorySearchNext();    AssertCurrentSubmission("100"); // NB: Reuses existing pattern.
-            _operations.HistorySearchNext();    AssertCurrentSubmission("4"); // Restoring input results in non-matching history entry after SearchNext.
-            _operations.HistoryNext();          AssertCurrentSubmission("4");
+            Window.Operations.HistorySearchNext();    AssertCurrentSubmission("10"); // Pattern is captured before this step.
+            Window.Operations.HistoryNext();          AssertCurrentSubmission("20"); // NB: Doesn't match pattern.
+            Window.Operations.HistorySearchNext();    AssertCurrentSubmission("100"); // NB: Reuses existing pattern.
+            Window.Operations.HistorySearchNext();    AssertCurrentSubmission("4"); // Restoring input results in non-matching history entry after SearchNext.
+            Window.Operations.HistoryNext();          AssertCurrentSubmission("4");
         }
 
         [WpfFact]
@@ -518,19 +551,19 @@ namespace Microsoft.VisualStudio.InteractiveWindow.UnitTests
             InsertAndExecuteInputs("1", "2", "10", "20", "30", "100", "200");
             SetActiveCode("4");
 
-            _operations.HistoryPrevious();  AssertCurrentSubmission("200");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("100");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("30");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("20");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("10");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("2");
-            _operations.HistoryPrevious();  AssertCurrentSubmission("1");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("200");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("100");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("30");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("20");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("10");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("2");
+            Window.Operations.HistoryPrevious();  AssertCurrentSubmission("1");
 
-            _operations.HistorySearchNext();    AssertCurrentSubmission("10"); // Pattern is captured before this step.
-            _operations.HistoryNext("2");       AssertCurrentSubmission("20"); // NB: Doesn't match pattern.
-            _operations.HistorySearchNext();    AssertCurrentSubmission("100"); // NB: Reuses existing pattern.
-            _operations.HistorySearchNext();    AssertCurrentSubmission("4"); // Restoring input results in non-matching history entry after SearchNext.
-            _operations.HistoryNext("2");       AssertCurrentSubmission("4");
+            Window.Operations.HistorySearchNext();    AssertCurrentSubmission("10"); // Pattern is captured before this step.
+            Window.Operations.HistoryNext("2");       AssertCurrentSubmission("20"); // NB: Doesn't match pattern.
+            Window.Operations.HistorySearchNext();    AssertCurrentSubmission("100"); // NB: Reuses existing pattern.
+            Window.Operations.HistorySearchNext();    AssertCurrentSubmission("4"); // Restoring input results in non-matching history entry after SearchNext.
+            Window.Operations.HistoryNext("2");       AssertCurrentSubmission("4");
         }
     }
 }
