+++ src/InteractiveWindow/EditorTest/InteractiveWindowTest.csproj
@@ -76,7 +76,9 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="InteractiveWindowClipboardTests.cs" />
     <Compile Include="InteractiveWindowEditorsFactoryService.cs" />
+    <Compile Include="InteractiveWindowTestBase.cs" />
     <Compile Include="InteractiveWindowTestHost.cs" />
     <Compile Include="HistoryTests.cs" />
     <Compile Include="InteractiveWindowHistoryTests.cs" />
